Relatório semestral e Conselho de classe - Botão de impressão só é liberado após selecionar um aluno
