+++ src/SME.SGP.WebClient/src/paginas/Fechamento/ConselhoClasse/conselhoClasse.js
@@ -1,30 +1,28 @@
-import React, { useEffect, useState, useCallback } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Loader } from '~/componentes';
+import AlertaModalidadeInfantil from '~/componentes-sgp/AlertaModalidadeInfantil/alertaModalidadeInfantil';
 import Cabecalho from '~/componentes-sgp/cabecalho';
 import Alert from '~/componentes/alert';
-import Button from '~/componentes/button';
 import Card from '~/componentes/card';
-import { Colors } from '~/componentes/colors';
+import RotasDto from '~/dtos/rotasDto';
 import {
+  limparDadosConselhoClasse,
   setAlunosConselhoClasse,
   setDadosAlunoObjectCard,
-  limparDadosConselhoClasse,
 } from '~/redux/modulos/conselhoClasse/actions';
-import { erros, erro, sucesso } from '~/servicos/alertas';
+import { erros } from '~/servicos/alertas';
 import ServicoConselhoClasse from '~/servicos/Paginas/ConselhoClasse/ServicoConselhoClasse';
+import { verificaSomenteConsulta } from '~/servicos/servico-navegacao';
+import { ehTurmaInfantil } from '~/servicos/Validacoes/validacoesInfatil';
 import { Container } from './conselhoClasse.css';
+import BotaoGerarRelatorioConselhoClasseTurma from './DadosConselhoClasse/BotaoGerarRelatorioConselhoClasseTurma/botaoGerarRelatorioConselhoClasseTurma';
 import BotaoOrdenarListaAlunos from './DadosConselhoClasse/BotaoOrdenarListaAlunos/botaoOrdenarListaAlunos';
 import BotoesAcoesConselhoClasse from './DadosConselhoClasse/BotoesAcoes/botoesAcoesConselhoClasse';
 import DadosConselhoClasse from './DadosConselhoClasse/dadosConselhoClasse';
 import ObjectCardConselhoClasse from './DadosConselhoClasse/ObjectCardConselhoClasse/objectCardConselhoClasse';
 import TabelaRetratilConselhoClasse from './DadosConselhoClasse/TabelaRetratilConselhoClasse/tabelaRetratilConselhoClasse';
 import servicoSalvarConselhoClasse from './servicoSalvarConselhoClasse';
-import AlertaModalidadeInfantil from '~/componentes-sgp/AlertaModalidadeInfantil/alertaModalidadeInfantil';
-import modalidade from '~/dtos/modalidade';
-import RotasDto from '~/dtos/rotasDto';
-import { verificaSomenteConsulta } from '~/servicos/servico-navegacao';
-import { ehTurmaInfantil } from '~/servicos/Validacoes/validacoesInfatil';
 
 const ConselhoClasse = () => {
   const dispatch = useDispatch();
@@ -35,26 +33,11 @@ const ConselhoClasse = () => {
   const permissoesTela = usuario.permissoes[RotasDto.CONSELHO_CLASSE];
 
   const [carregandoGeral, setCarregandoGeral] = useState(false);
-  const [imprimindo, setImprimindo] = useState(false);
   const [exibirListas, setExibirListas] = useState(false);
-  const [podeImprimir, setPodeImprimir] = useState(false);
-
-  const conselhoClasseId = useSelector(
-    store => store.conselhoClasse.dadosPrincipaisConselhoClasse.conselhoClasseId
-  );
-
-  const fechamentoTurmaId = useSelector(
-    store =>
-      store.conselhoClasse.dadosPrincipaisConselhoClasse.fechamentoTurmaId
-  );
 
   const modalidadesFiltroPrincipal = useSelector(
     store => store.filtro.modalidades
   );
-
-  useEffect(() => {
-    setPodeImprimir(conselhoClasseId);
-  }, [conselhoClasseId]);
 
   const obterListaAlunos = useCallback(async () => {
     setCarregandoGeral(true);
@@ -129,21 +112,6 @@ const ConselhoClasse = () => {
     return false;
   };
 
-  const gerarConselhoClasseTurma = async () => {
-    setImprimindo(true);
-    await ServicoConselhoClasse.gerarConselhoClasseTurma(
-      conselhoClasseId,
-      fechamentoTurmaId
-    )
-      .then(() => {
-        sucesso(
-          'Solicitação de geração do relatório gerada com sucesso. Em breve você receberá uma notificação com o resultado.'
-        );
-      })
-      .finally(setImprimindo(false))
-      .catch(e => erro(e));
-  };
-
   return (
     <Container>
       {!turmaSelecionada.turma &&
@@ -182,17 +150,7 @@ const ConselhoClasse = () => {
                 <>
                   <div className="col-md-12 mb-2 d-flex">
                     <BotaoOrdenarListaAlunos />
-                    <Loader loading={imprimindo}>
-                      <Button
-                        className="btn-imprimir"
-                        icon="print"
-                        color={Colors.Azul}
-                        border
-                        onClick={() => gerarConselhoClasseTurma()}
-                        disabled={!podeImprimir}
-                        id="btn-imprimir-relatorio-pendencias"
-                      />
-                    </Loader>
+                    <BotaoGerarRelatorioConselhoClasseTurma />
                   </div>
                   <div className="col-md-12 mb-2">
                     <TabelaRetratilConselhoClasse
@@ -200,10 +158,7 @@ const ConselhoClasse = () => {
                       permiteOnChangeAluno={permiteOnChangeAluno}
                     >
                       <>
-                        <ObjectCardConselhoClasse
-                          conselhoClasseId={conselhoClasseId}
-                          fechamentoTurmaId={fechamentoTurmaId}
-                        />
+                        <ObjectCardConselhoClasse />
                         <DadosConselhoClasse
                           turmaCodigo={turmaSelecionada.turma}
                           modalidade={turmaSelecionada.modalidade}
